SmartGraderCore/requirements.txt
mysqlclient
asn1crypto
blinker
cryptography
defusedxml
django
django-debug-toolbar
djangorestframework
oauthlib
olefile
pillow
pip
pycparser
pygments
pyjwt
pyopenssl
pysocks
python-memcached
python3-openid
requests
requests-oauthlib
setuptools
simplejson
six
sqlparse
urllib3
wheel
django-safedelete
pdf2image
gunicorn
Keras==2.3.1
Keras-Applications
Keras-Preprocessing
PyPDF2
numpy
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
pandas
scikit-image
scikit-learn
scipy
tensorflow==2.2.0
tensorflow-estimator==2.2.0
